opcodes: Add bfd handles to disassemble_info
As `abfd` and `obfd` handles are useful in multiple places, it's
reasonable to add them in `struct disassemble_info`. The current bfd
metadata stays as is, as bfd handles are not available in all contexts
(and don't need to be).

This change enables more unified disassembler function derivation,
moving the optional `abfd` handling to `print_insn_*` functions.

The `obfd` handle is a function (`get_obfd_for_addr`), which resolves
the object file bfd depending on the disassembled address. This enables
more granular information for the disassemblers, especially if the ELF
contains arch-specific information which can be used to provide better
disassembly output. A notable example is RISC-V, which can use shared
libraries built for a different subset of ISA extensions.

Signed-off-by: Marek Pikuła <m.pikula@partner.samsung.com>