RISC-V: Use obfd for disassembly instead of abfd
In case the disassembled ELF is built for a different subset of ISA
extensions than the target binary (abfd) the disassembly might be
different.

For example, a binary might be built for a general `rv64gc` target, but
a library that the binary loads might be built for `rv64gcv`. In that
case, when only abfd provides information to disassembler, the vector
instructions don't get disassembled correctly (and are shown as generic
`.insn`). With this change, the disassembler checks which object file is
in scope. Thus, it adjusts the output to the ISA subset present in
`Tag_RISCV_arch` from the object file in the current scope.

Signed-off-by: Marek Pikuła <m.pikula@partner.samsung.com>